move all bottles to subapps
