+++ jumpscale/packages/admin/bottle/admin.py
@@ -82,4 +82,4 @@ def export():
     )
 
 
-app = SessionMiddleware(app, SESSION_OPTS)
+# app = SessionMiddleware(app, SESSION_OPTS)
