+++ jumpscale/packages/threebot_deployer/bottle/solutions.py
@@ -119,4 +119,4 @@ def accept():
         )
 
 
-app = SessionMiddleware(app, SESSION_OPTS)
+# app = SessionMiddleware(app, SESSION_OPTS)
