+++ jumpscale/packages/vdc_dashboard/bottle/export.py
@@ -21,4 +21,4 @@ def export():
     )
 
 
-app = SessionMiddleware(app, SESSION_OPTS)
+# app = SessionMiddleware(app, SESSION_OPTS)
